+++ src/main/java/com/newer/sc/common/entity/Teacher.java
@@ -1,6 +1,5 @@
 package com.newer.sc.common.entity;
 
-import java.math.BigInteger;
 import java.sql.Date;
 
 import org.springframework.stereotype.Component;
@@ -13,7 +12,7 @@ import org.springframework.stereotype.Component;
 @Component
 public class Teacher {
 	
-	BigInteger tid;
+	Long tid;
 	
 	String tname;
 	
@@ -29,11 +28,11 @@ public class Teacher {
 	
 	College college;
 
-	public BigInteger getTid() {
+	public Long getTid() {
 		return tid;
 	}
 
-	public void setTid(BigInteger tid) {
+	public void setTid(Long tid) {
 		this.tid = tid;
 	}
